Split presentation UI drawing
